Consolidate lowering behind snail-python
