+++ crates/snail-ast/README.md
@@ -24,7 +24,7 @@ None - this crate has no dependencies and provides pure data structures.
 - **snail-error**: Uses `SourceSpan` for error reporting with source locations
 - **snail-parser**: Produces `Program` and `AwkProgram` as output
 - **snail-lower**: Consumes Snail AST and transforms it to Python `ast` nodes via pyo3
-- **snail-core**: Re-exports all types for the unified API
+- **snail-python**: Uses AST types as part of the compilation pipeline
 
 ## Design
