--- crates/snail-core/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "snail-core"
-version = "0.6.3"
-edition.workspace = true
-
-[dependencies]
-snail-ast = { path = "../snail-ast" }
-snail-error = { path = "../snail-error" }
-snail-parser = { path = "../snail-parser" }
-snail-lower = { path = "../snail-lower" }
-pyo3 = { version = "0.21", features = ["abi3-py310"] }
