+++ crates/snail-lower/src/lib.rs
@@ -1,20 +1,5 @@
-// Module declarations
-mod awk;
-mod constants;
-mod expr;
-mod helpers;
-mod map;
-mod operators;
-mod program;
-mod py_ast;
-mod stmt;
-
-// Re-export public API
-pub use constants::*;
-pub use map::{
-    lower_map_program, lower_map_program_with_auto_print, lower_map_program_with_begin_end,
-};
-pub use program::{
-    lower_awk_program, lower_awk_program_with_auto_print, lower_program,
+pub use snail_python::{
+    PyObject, Python, lower_awk_program, lower_awk_program_with_auto_print, lower_map_program,
+    lower_map_program_with_auto_print, lower_map_program_with_begin_end, lower_program,
     lower_program_with_auto_print,
 };
