+++ crates/snail-proptest/Cargo.toml
@@ -7,13 +7,11 @@ edition = "2024"
 snail-ast = { path = "../snail-ast" }
 snail-parser = { path = "../snail-parser" }
 snail-lower = { path = "../snail-lower" }
-snail-core = { path = "../snail-core" }
 snail-error = { path = "../snail-error" }
 proptest = "1.5"
 
 [dev-dependencies]
 tempfile = "3"
-pyo3 = { version = "0.21", features = ["abi3-py310"] }
 
 [features]
 run-proptests = []
