+++ crates/snail-proptest/tests/snail_features.rs
@@ -1,7 +1,7 @@
 #![cfg(feature = "run-proptests")]
 
 use proptest::prelude::*;
-use pyo3::prelude::*;
+use snail_lower::{PyObject, Python};
 use snail_proptest::arbitrary::*;
 use snail_proptest::utils::*;
